scripts/screenshot/requirements.txt
freetype-py==2.2.0
